更新界面

+++ app/src/main/java/com/omarea/library/device/BatteryCapacity.java
@@ -17,8 +17,7 @@ public class BatteryCapacity {
 
         try {
             mPowerProfile = Class.forName(POWER_PROFILE_CLASS).getConstructor(Context.class).newInstance(context);
-            batteryCapacity = (double) Class.forName(POWER_PROFILE_CLASS).getMethod("getBatteryCapacity").invoke(mPowerProfile);
-
+            batteryCapacity = (int) ((double) Class.forName(POWER_PROFILE_CLASS).getMethod("getBatteryCapacity").invoke(mPowerProfile));
         } catch (Exception e) {
             e.printStackTrace();
         }
