+++ app/src/main/java/com/omarea/model/BatteryStatus.java
@@ -3,7 +3,7 @@ package com.omarea.model;
 public class BatteryStatus {
     public long time;
     public int level = -1;
-    public float temperature = -1;
+    public double temperature = -1;
     public int status = 0;
     public String statusText;
     public int io = -1;
